requirements.txt
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0
chromadb>=0.4.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
Pillow>=10.0.0
numpy>=1.24.0
click>=8.1.0
tqdm>=4.65.0


